description:new updates

+++ exp_2_without/pages.py
@@ -293,9 +293,9 @@ class Final(Page):
         return self.player.participant.vars['consent'].lower() == 'consent'
 
     def vars_for_template(self):
-        self.player.payoff_cem = round(self.player.participant.vars['payoff_cem']*0.02,2)
-        self.player.payoff_trust = round(self.player.participant.vars['payoff_trust']*0.02,2)
-        self.player.payoff_total = round(self.player.total_bonus+(self.player.participant.vars['payoff_trust']+self.player.participant.vars['payoff_cem'])*0.02,)
+        self.player.payoff_cem = round(float(self.player.participant.vars['payoff_cem']*0.02),2)
+        self.player.payoff_trust = round(float(self.player.participant.vars['payoff_trust']*0.02),2)
+        self.player.payoff_total = round(float(self.player.total_bonus+(self.player.participant.vars['payoff_trust']+self.player.participant.vars['payoff_cem'])*0.02),)
 
         return {'id': self.player.id_in_group,
                 'payoff_trust':self.player.payoff_trust,
